Renamed asgard-client --> asgard-service
