+++ lib/asgard-service.js
