+++ lib/deployer.js
@@ -1,8 +1,8 @@
 const internals = {};
 
-export function create(asgard_client) {
+export function create(asgard_service) {
   return {
-    startDeploymentInCluster: internals.startDeploymentInCluster.bind(null, asgard_client),
+    startDeploymentInCluster: internals.startDeploymentInCluster.bind(null, asgard_service),
   };
 }
